requirements.txt
python-dotenv
pymongo
discord.py
loguru
requests
